Separate Message to InfoMessage and ErrorMessage
